feature: Add UI for displaying messages
ChatScreenState definition was modified to include TickerProviderStateMixin for serving as vsync argument in preventing offscreen animations when initializing an AnimationController

ChatMessage was modified to store a reference to an AnimationController

ChatScreenState.handleSubmitted() was modified to instantiate, attach, and play forward the AnimationController on submission of each ChatMessage.

ChatMessage was modified to wrap each high-level message Container with a SizeTransition that generates an easeOut CurvedAnimation effect.

ChatScreenSize destructor was modified to dispose of each ChatMessage AnimationController before disposing of the screen, refactoring the foreach loop with method chaining.